server/ml/requirements.txt
ultralytics==8.3.169
Pillow==11.3.0
